tests/exports: add tests for comment and item exports and restructure
